Moved the coordinate interfaces to the jsettlers.common.utils.coordinates package
