+++ jsettlers.common/src/main/java/jsettlers/common/utils/coordinates/ICoordinateFunction.java
@@ -12,11 +12,11 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package jsettlers.common.utils.interfaces;
+package jsettlers.common.utils.coordinates;
 
 /**
  * Created by Andreas Eberle on 06.01.2017.
  */
-public interface ICoordinateWithRadiusFunction<T> {
-	T apply(int x, int y, int radius);
+public interface ICoordinateFunction<T> {
+	T apply(int x, int y);
 }
